+++ Modules/Intent.Modules.Application.DomainInteractions/InteractionStrategies/ODataQueryInteractionStrategy.cs
@@ -16,7 +16,8 @@ public class ODataQueryInteractionStrategy : IInteractionStrategy
     private const string ODataQueryStereoType = "ODataQuery"; // Note that IDs should work too.
     public bool IsMatch(IElement interaction)
     {
-        return interaction.IsQueryEntityActionTargetEndModel() && interaction.AsQueryEntityActionTargetEndModel().OtherEnd().Element.HasStereotype(ODataQueryStereoType);
+        return interaction.IsQueryEntityActionTargetEndModel() &&
+               interaction.AsQueryEntityActionTargetEndModel().OtherEnd().Element.HasStereotype(ODataQueryStereoType);
     }
 
     public void ImplementInteraction(ICSharpClassMethodDeclaration method, IElement interactionElement)