+++ Modules/Intent.Modules.Application.MediatR.CRUD/FactoryExtensions/MongoDbRepositoryUpdateExtension.cs
@@ -8,7 +8,6 @@ using Intent.Modules.Common.CSharp.Interactions;
 using Intent.Modules.Common.CSharp.Templates;
 using Intent.Modules.Common.Plugins;
 using Intent.Modules.Common.Templates;
-using Intent.Modules.Eventing.Contracts.InteractionStrategies;
 using Intent.Plugins.FactoryExtensions;
 using Intent.RoslynWeaver.Attributes;